pyproject.toml
[tool.poetry]
name = "papzi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sv-ttk = "^2.6.0"
pillow = "^10.3.0"

datasets = "^2.19.1"
torch = [ 
{ platform = "darwin", version = "^2.3.0", source= "pypi" },
{ platform="win32", version = "^2.3.0+cu121", source = "pytorch-gpu"},
]
torchvision = [
{platform = "darwin", version = "^0.18.0", source= "pypi" },
{platform="win32", version = "^0.18.0+cu121", source = "pytorch-gpu"}
]

torchaudio = [
{platform = "darwin", version = "^2.3.0", source= "pypi" },
{platform="win32", version = "^2.3.0+cu121", source = "pytorch-gpu"}
]

opencv-python = "^4.9.0.80"
opencv-contrib-python = "^4.10.0.82"
numpy = "1.26.4"



[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"
pip = "^24.0"
install = "^1.3.5"
ipywidgets = "^8.1.3"
selenium = "^4.21.0"
webdriver-manager = "^4.0.1"
aiohttp = "^3.9.5"
undetected-chromedriver = "^3.5.5"

[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[tool.black]
line-length = 79  # Or any value you prefer

[tool.isort]
profile = "black"

[tool.poetry.scripts]
trainn = 'papzi.train_model:main'
papzi = 'papzi_gui.app:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

